Updates from smoke test
The READMEs on two samples were updated to provide necessary information. In one case for the PHP sample to describe a Dev Environment to help the end user. In the case of the SQLite3 sample to indicate that the NodeJS v13 or later works as it was tested to work on NodeJS v18.